requirements.txt
Flask
torch
transformers
Pillow
wikipedia
